requirements.txt
Flask==3.0.1
Flask-HTTPAuth==4.8.0
PyYAML==6.0.1
Werkzeug==3.0.1
